refactor(front): move file_system entry types into types/api
The file_system/types.ts module mixes pure file-system entry types + their DTOs
with runtime exports (a class, error guard, consts, helpers). Move only the pure
type tree to types/api; leave all runtime in lib.

- Moved to types/api/file_system/types.ts: FileSystemEntryBase, FileSystemEntry,
  FileSystemFileEntry, FileSystemDirectoryEntry, GetSpaceFilesResponseBody,
  PostSpaceFolderResponseBody.
- Kept in lib: FileSystemMount(+Kind), DustFileSystemError(+code+guard),
  SCOPED_/LEGACY_PREFIX consts, conversationScopedPath/podScopedPath.
- 20 consumers + the barrel updated (split-aware: types -> @app/types, runtime ->
  @app/lib). Also resolves the deferred Pass B file conversation/files.ts
  (FileSystemEntry now from @app/types).
