+++ front/components/file_explorer/FileExplorer.tsx
@@ -25,7 +25,7 @@ import {
   getScopedRelativePath,
   isFileExplorerMovableFile,
 } from "@app/components/file_explorer/utils";
-import type { FileSystemEntry } from "@app/lib/api/file_system/types";
+import type { FileSystemEntry } from "@app/types/api/file_system/types";
 import { isInteractiveContentType } from "@app/types/files";
 import { Err, type Result } from "@app/types/shared/result";
 import { cn, Edit04, FolderOpen, Trash01 } from "@dust-tt/sparkle";
