+++ front/components/file_explorer/utils.test.ts
@@ -9,7 +9,7 @@ import {
   getChildrenAtFolderPath,
   isFileExplorerMovableFile,
 } from "@app/components/file_explorer/utils";
-import type { FileSystemEntry } from "@app/lib/api/file_system/types";
+import type { FileSystemEntry } from "@app/types/api/file_system/types";
 import { frameContentType, frameSlideshowContentType } from "@app/types/files";
 import { describe, expect, it } from "vitest";
