+++ front/lib/api/assistant/conversation/files.ts
@@ -1,4 +1,4 @@
-import type { FileSystemEntry } from "@app/lib/api/file_system/types";
+import type { FileSystemEntry } from "@app/types/api/file_system/types";
 
 export type GetConversationFilesResponseBody = {
   files: FileSystemEntry[];
