+++ front/lib/api/file_system/index.ts
@@ -6,7 +6,6 @@ export {
 } from "@app/lib/api/file_system/dust_file_system";
 export type {
   DustFileSystemErrorCode,
-  FileSystemEntry,
   FileSystemMount,
 } from "@app/lib/api/file_system/types";
 export {
@@ -15,3 +14,4 @@ export {
   SCOPED_PREFIX_CONVERSATION,
   SCOPED_PREFIX_POD,
 } from "@app/lib/api/file_system/types";
+export type { FileSystemEntry } from "@app/types/api/file_system/types";
