+++ front/lib/api/files/file_system_ops.ts
@@ -7,7 +7,6 @@
 import type { DustFileSystem } from "@app/lib/api/file_system";
 import {
   DustFileSystemError,
-  type FileSystemEntry,
   SCOPED_PREFIX_CONVERSATION,
   SCOPED_PREFIX_POD,
 } from "@app/lib/api/file_system/types";
@@ -15,6 +14,7 @@ import { decodeBuffer } from "@app/lib/api/files/utils";
 import type { Authenticator } from "@app/lib/auth";
 import { FileResource } from "@app/lib/resources/file_resource";
 import logger from "@app/logger/logger";
+import type { FileSystemEntry } from "@app/types/api/file_system/types";
 import type { FileUseCase, FileUseCaseMetadata } from "@app/types/files";
 import {
   contentTypeFromFileName,
