+++ front/lib/api/mcp_server/tools/files/list_output.ts
@@ -1,11 +1,11 @@
 import type { DustFileSystem } from "@app/lib/api/file_system";
-import type {
-  FileSystemEntry,
-  FileSystemFileEntry,
-} from "@app/lib/api/file_system/types";
 import { enrichListWithFileResourceIds } from "@app/lib/api/files/file_system_ops";
 import { parseProcessedFilename } from "@app/lib/api/files/mount_path";
 import type { Authenticator } from "@app/lib/auth";
+import type {
+  FileSystemEntry,
+  FileSystemFileEntry,
+} from "@app/types/api/file_system/types";
 import {
   isInteractiveContentType,
   stripMimeParameters,
